moved output helpers into an output_formatters directory
