+++ lib/napa.rb
@@ -18,9 +18,9 @@ require 'napa/stats'
 require 'napa/active_record_extensions/filter_by_hash'
 require 'napa/grape_extensions/error_formatter'
 require 'napa/grape_extensions/grape_helpers'
-require 'napa/grape_extensions/entity'
-require 'napa/grape_extensions/include_nil'
-require 'napa/grape_extensions/representer'
+require 'napa/output_formatters/entity'
+require 'napa/output_formatters/include_nil'
+require 'napa/output_formatters/representer'
 require 'napa/grape_extenders'
 require 'napa/middleware/logger'
 require 'napa/middleware/app_monitor'
