+++ lib/napa/output_formatters/entity.rb
