+++ lib/napa/output_formatters/include_nil.rb
